Moving FirstUniqueWindowTestCase.java, UniqueExternalTimeBatchWindowTestCase.java, UniqueWindowTestCase.java from siddhi core to siddhi-execution-unique repository.
