+++ component/src/test/java/org/wso2/extension/siddhi/execution/unique/UniqueExternalTimeBatchWindowTestCase.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.siddhi.core.query.window;
+package org.wso2.extension.siddhi.execution.unique;
 
 import org.apache.log4j.Logger;
 import org.junit.Assert;
